refactor(core): share locale resolution and dedupe org subscription query
Follow-up cleanup for the Better Auth hooks/email port:

- Move the locale constants, Accept-Language resolver, and email-locale
  helpers into @sokosumi/utils so core and web share a single source of
  truth instead of byte-for-byte copies. Delete the duplicated core
  (lib/i18n, lib/email-locale) and web (i18n/locale-resolution) modules
  and repoint every importer at the shared package. Web keeps only its
  UI-specific locale extras (auto-detect, localStorage key, cookie
  serializers).
- Add resolveOrganizationBillingPlanWithActiveSubscription, which returns
  the active subscription it already fetched, so the seat/credit sync and
  invitation precheck no longer query the active subscription twice on the
  self-serve path. The existing resolveOrganizationBillingPlan delegates to
  it, so its other callers are unchanged.
