+++ packages/database/src/helpers/organization-billing-plan.ts
@@ -5,6 +5,7 @@ import {
 import {
   EnterpriseContractStatus,
   type Prisma,
+  type Subscription,
 } from "../generated/prisma/client.js";
 import { subscriptionRepository } from "../repositories/subscription.repository.js";
 import {
@@ -40,6 +41,33 @@ export async function resolveOrganizationBillingPlan(
   tx: Prisma.TransactionClient,
   now: Date = new Date(),
 ): Promise<OrganizationBillingPlan> {
+  const { billingPlan } =
+    await resolveOrganizationBillingPlanWithActiveSubscription(
+      organizationId,
+      tx,
+      now,
+    );
+
+  return billingPlan;
+}
+
+/**
+ * Resolves the billing plan and returns the active subscription it fetched.
+ *
+ * Callers that also need the active subscription (e.g. seat/credit syncing)
+ * should use this instead of calling `resolveOrganizationBillingPlan` followed
+ * by a separate subscription lookup, which would query the same row twice.
+ * For enterprise-contract organizations no subscription is fetched and
+ * `activeSubscription` is `null`.
+ */
+export async function resolveOrganizationBillingPlanWithActiveSubscription(
+  organizationId: string,
+  tx: Prisma.TransactionClient,
+  now: Date = new Date(),
+): Promise<{
+  billingPlan: OrganizationBillingPlan;
+  activeSubscription: Subscription | null;
+}> {
   const activeContract = await tx.enterpriseContract.findFirst({
     where: {
       organizationId,
@@ -60,18 +88,21 @@ export async function resolveOrganizationBillingPlan(
     });
 
     return {
-      mode: "enterprise_contract",
-      plan: "enterprise",
-      isConsumable,
-      purchasedSeats: activeContract.seats,
-      contractId: activeContract.id,
-      endsAt: deriveEnterpriseContractEndDate(
-        activeContract.activatedAt,
-        activeContract.periodCount,
-      ),
-      activatedAt: activeContract.activatedAt,
-      cancelAtPeriodEnd: false,
-      periodEnd: null,
+      billingPlan: {
+        mode: "enterprise_contract",
+        plan: "enterprise",
+        isConsumable,
+        purchasedSeats: activeContract.seats,
+        contractId: activeContract.id,
+        endsAt: deriveEnterpriseContractEndDate(
+          activeContract.activatedAt,
+          activeContract.periodCount,
+        ),
+        activatedAt: activeContract.activatedAt,
+        cancelAtPeriodEnd: false,
+        periodEnd: null,
+      },
+      activeSubscription: null,
     };
   }
 
@@ -90,11 +121,14 @@ export async function resolveOrganizationBillingPlan(
     plan === "free" ? 0 : resolvePurchasedSeats(subscription?.seats);
 
   return {
-    mode: "self_serve",
-    plan,
-    purchasedSeats,
-    subscriptionId: subscription?.id ?? null,
-    cancelAtPeriodEnd: subscription?.cancelAtPeriodEnd ?? false,
-    periodEnd: subscription?.periodEnd ?? null,
+    billingPlan: {
+      mode: "self_serve",
+      plan,
+      purchasedSeats,
+      subscriptionId: subscription?.id ?? null,
+      cancelAtPeriodEnd: subscription?.cancelAtPeriodEnd ?? false,
+      periodEnd: subscription?.periodEnd ?? null,
+    },
+    activeSubscription: subscription,
   };
 }
